+++ src/main/java/com/chatspring/chatspring/stock/statistics/TradeStatisticsService.java
@@ -25,9 +25,9 @@ public class TradeStatisticsService {
         tradeStatistics.setCountSellPrice2((Integer) statistics.get("count_sell_price_2"));
         tradeStatistics.setCountSellPrice3((Integer) statistics.get("count_sell_price_3"));
         tradeStatistics.setAvgReachTime((String) statistics.get("avg_reach_time"));
-        tradeStatistics.setWinRatioKospi((Integer) statistics.get("win_ratio_kospi"));
-        tradeStatistics.setWinRatioKosdaq((Integer) statistics.get("win_ratio_kosdaq"));
-        tradeStatistics.setRatioMaxBuy((Integer) statistics.get("ratio_max_buy"));
+        tradeStatistics.setWinRatioKospi((Float) statistics.get("win_ratio_kospi"));
+        tradeStatistics.setWinRatioKosdaq((Float) statistics.get("win_ratio_kosdaq"));
+        tradeStatistics.setRatioMaxBuy((Float) statistics.get("ratio_max_buy"));
 
         tradeStatisticsRepository.save(tradeStatistics);
     }
